feat: style cm info (#145)
