+++ components/forms/index.ts
@@ -1,2 +1,2 @@
-export { default as UpdateCreateCandyMachineForm } from './UpdateCreateCandyMachineForm'
+export { default as UpdateCandyMachine } from './UpdateCandyMachine'
 export { default as CreateCandyMachine } from './CreateCandyMachine'
