+++ pages/[id].tsx
@@ -1,16 +1,7 @@
 import { AnchorProvider, BN, Program } from '@project-serum/anchor'
 import { useAnchorWallet } from '@solana/wallet-adapter-react'
-import { PublicKey } from '@solana/web3.js'
-import {
-    Spinner,
-    Title,
-    UpdateCreateCandyMachineForm,
-    Carousel,
-    ExplorerLinks,
-    NftCard,
-    RefreshButton,
-} from 'components'
-import { useRPC } from 'hooks'
+import { Connection, PublicKey } from '@solana/web3.js'
+import { Title, UpdateCandyMachine, Carousel, NftCard, RefreshButton } from 'components'
 import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/constants'
 import { IFetchedCandyMachineConfig } from 'lib/candy-machine/interfaces'
 import { Nft } from 'lib/nft/interfaces'
@@ -21,7 +12,8 @@ import { useEffect, useState } from 'react'
 import { useSetRecoilState } from 'recoil'
 import { nftsState } from 'lib/recoil-store/atoms'
 import { getAllNftsByCM, getNftByMint } from 'lib/nft/actions'
-import { useMintCandyMachine } from 'hooks'
+import { Button, Spinner } from '@primer/react'
+import { useMintCandyMachine, useRPC } from 'hooks'
 
 const CandyMachine: NextPage = () => {
     const router = useRouter()
@@ -159,249 +151,218 @@ const CandyMachine: NextPage = () => {
                 <meta name='description' content='Generated by create next app' />
                 <link rel='icon' href='/favicon.ico' />
             </Head>
-            <div className='d-flex flex-justify-center flex-items-center flex-column'>
-                <Title text='Update Candy Machine' />
-                <div className='mt-8 d-flex flex-column text-center'>
-                    <span className='break-all border border-slate-300 shadow-xl py-2 px-4 rounded-lg text-center'>
-                        {candyMachineAccount}{' '}
-                    </span>
-                    <ExplorerLinks
-                        type='account'
-                        value={candyMachineAccount as string}
-                        connection={connection}
-                        text={'View'}
-                    />
-                </div>
-                {isLoading && <Spinner />}
+            <div className='d-flex flex-justify-center flex-column'>
+                <h2 className='my-5 wb-break-all'>[CM]: {candyMachineAccount}</h2>
+                {isLoading && (
+                    <div className='d-flex flex-items-center'>
+                        <h3 className='mr-3'>Loading...</h3>
+                        <Spinner size='small' />
+                    </div>
+                )}
                 {error && (
-                    <div className='d-flex flex-column items-center justify-center mt-11'>
-                        {error}
-                        <button className='rounded-lg bg-slate-400 p-2 mt-4' onClick={() => fetchCandyMachine()}>
+                    <div className='d-flex flex-column items-center justify-center my-5 col-12 col-md-8 col-lg-6'>
+                        <h3 className='color-fg-accent'> Error fetching candy machine config</h3>
+                        <Button
+                            className='rounded-lg bg-slate-400 p-2 mt-4'
+                            onClick={() => fetchCandyMachine()}
+                            sx={{ width: 'fit-content' }}
+                        >
                             Fetch again
-                        </button>
+                        </Button>
                     </div>
                 )}
 
-                {!error && candyMachineConfig?.uuid && (
-                    <div className='mt-5 d-flex flex-column text-center'>
-                        <span className='mt-5'>
-                            There are{' '}
-                            {new BN(candyMachineConfig.itemsAvailable).toNumber() -
-                                new BN(candyMachineConfig.itemsRedeemed).toNumber()}{' '}
-                            unminted NFT.
-                        </span>
-                        <span className='mt-2'>
-                            {new BN(candyMachineConfig.itemsRedeemed).toNumber()} redeemed NFT.
-                        </span>
-                        {new BN(candyMachineConfig.itemsAvailable).toNumber() -
-                            new BN(candyMachineConfig.itemsRedeemed).toNumber() !==
-                            0 && (
-                            <>
-                                <label
-                                    htmlFor='cache'
-                                    className='m-8 px-4 py-2 rounded-xl bg-slate-200 hover:bg-slate-100 transition-all duration-300 ease-linear font-medium border border-gray-500 inline-block cursor-pointer'
-                                >
-                                    Upload Cache file to view NFTs
-                                </label>
-                                <input
-                                    type='file'
-                                    id='cache'
-                                    name='cache'
-                                    className='w-full px-2 hidden'
-                                    onChange={viewNfts}
-                                />
-                            </>
-                        )}
-                        {nfts.length !== 0 && (
-                            <Carousel
-                                carouselData={nfts.map((nft) => ({
-                                    title: nft.name,
-                                    image: nft.image,
-                                }))}
+                {!error && !isLoading && candyMachineConfig?.uuid && (
+                    <div className='d-flex flex-column'>
+                        <div className='d-flex flex-column mb-6'>
+                            <h3>Configuration</h3>
+                            <div className='border-y width-full my-4' />
+                            <UpdateCandyMachine
+                                fetchedValues={candyMachineConfig}
+                                candyMachinePubkey={candyMachineAccount}
                             />
-                        )}
-                        <UpdateCreateCandyMachineForm
-                            fetchedValues={candyMachineConfig}
-                            updateCandyMachine
-                            candyMachinePubkey={candyMachineAccount}
-                        />
-                    </div>
-                )}
-                <div className='d-flex flex-justify-center flex-items-start flex-column mb-10 width-full p-0'>
-                    <div className='d-flex flex-justify-between flex-items-center width-full mb-5'>
-                        <h3 className='r-0'>NFTs</h3>
-                        <div className='l-0 d-flex flex-justify-end flex-items-center'>
-                            <span className='pr-2'>5/10 Minted</span>
-                            <RefreshButton onClick={refreshNfts} isLoading={isReloading} />
                         </div>
                     </div>
-                    <div className='border-y width-full' />
-                    {hasCollection && (
+                )}
+
+                {!isLoading && (
+                    <div className='d-flex flex-justify-center flex-items-start flex-column mb-10 width-full p-0'>
+                        <div className='d-flex flex-justify-between flex-items-center width-full mb-5'>
+                            <h3 className='r-0'>NFTs</h3>
+                            <div className='l-0 d-flex flex-justify-end flex-items-center'>
+                                <span className='pr-2'>5/10 Minted</span>
+                                <RefreshButton onClick={refreshNfts} isLoading={isReloading} />
+                            </div>
+                        </div>
+                        <div className='border-y width-full' />
+                        {hasCollection && (
+                            <div className='mt-5'>
+                                <h4>Collection</h4>
+
+                                {collectionNft && (
+                                    <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
+                                        <NftCard
+                                            title={collectionNft.name}
+                                            imageLink={collectionNft.image}
+                                            buttons={[
+                                                {
+                                                    text: 'View in Solscan',
+                                                    as: 'link',
+                                                    variant: 'invisible',
+                                                    hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                                },
+                                            ]}
+                                        />
+                                    </div>
+                                )}
+                            </div>
+                        )}
                         <div className='mt-5'>
-                            <h4>Collection</h4>
-                            {collectionNft && (
-                                <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
+                            <h4>Minted NFTs - 5</h4>
+                            <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
+                                {itemsRemaining > 0 && (
                                     <NftCard
-                                        title={collectionNft.name}
-                                        imageLink={collectionNft.image}
+                                        title={'New NFT'}
                                         buttons={[
                                             {
-                                                text: 'View in Solscan',
-                                                as: 'link',
-                                                variant: 'invisible',
-                                                hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                                text: 'Mint 1 NFT',
+                                                isLoading: isUserMinting,
+                                                as: 'button',
+                                                variant: 'primary',
+                                                onClick: () => mintAccount(),
                                             },
                                         ]}
                                     />
-                                </div>
-                            )}
-                        </div>
-                    )}
-                    <div className='mt-5'>
-                        <h4>Minted NFTs - 5</h4>
-                        <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
-                            {itemsRemaining > 0 && (
+                                )}
                                 <NftCard
-                                    title={'New NFT'}
+                                    title={'CryptoDude #1'}
+                                    imageLink={'/favicon.ico'}
                                     buttons={[
                                         {
-                                            text: 'Mint 1 NFT',
-                                            isLoading: isUserMinting,
-                                            as: 'button',
-                                            variant: 'primary',
-                                            onClick: () => mintAccount(),
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
                                         },
                                     ]}
                                 />
-                            )}
-                            <NftCard
-                                title={'CryptoDude #1'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #2'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #3'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #4'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #5'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
+                                <NftCard
+                                    title={'CryptoDude #2'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #3'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #4'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #5'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                            </div>
+                        </div>
+                        <div className='mt-5'>
+                            <h4>Unminted NFTs - 10</h4>
+                            <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
+                                <NftCard
+                                    title={'CryptoDude #1'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #2'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #3'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #4'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                                <NftCard
+                                    title={'CryptoDude #5'}
+                                    imageLink={'/favicon.ico'}
+                                    buttons={[
+                                        {
+                                            text: 'View in Solscan',
+                                            as: 'link',
+                                            variant: 'invisible',
+                                            hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
+                                        },
+                                    ]}
+                                />
+                            </div>
                         </div>
                     </div>
-                    <div className='mt-5'>
-                        <h4>Unminted NFTs - 10</h4>
-                        <div className='d-flex flex-justify-start flex-items-center gap-5 mt-3'>
-                            <NftCard
-                                title={'CryptoDude #1'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #2'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #3'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #4'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                            <NftCard
-                                title={'CryptoDude #5'}
-                                imageLink={'/favicon.ico'}
-                                buttons={[
-                                    {
-                                        text: 'View in Solscan',
-                                        as: 'link',
-                                        variant: 'invisible',
-                                        hash: '14eoYMYLY19gtfE1gwWDhnjDD3fDjGTQTGyicBKT33Ns',
-                                    },
-                                ]}
-                            />
-                        </div>
-                    </div>
-                </div>
+                )}
             </div>
         </>
     )
